src/Services/Fraud/Tigerspike.Solv.Services.Fraud/Rules/SuspiciousCsatRule.cs
using System;
using System.Threading.Tasks;
using Tigerspike.Solv.Messaging.Fraud;

namespace Tigerspike.Solv.Services.Fraud.Rules
{
	public class SuspiciousCsatRule : IRule
	{
		/// <inheritdoc />
		public IDetectFraudCommand Message { get; set; }

		/// <inheritdoc />
		public Guid RuleId { get; set; }

		public SuspiciousCsatRule(IDetectFraudCommand message, Guid ruleId)
		{
			Message = message;
			RuleId = ruleId;
		}

		public Task<bool> IsMatchAsync()
		{
			return Task.FromResult(Message.Csat.Value == 5);
		}

		public bool IsValid()
		{
			return Message.Csat.HasValue;
		}
	}
}